requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
matplotlib>=3.6.0
pandas>=2.0.0
satpy>=0.45.0
requests>=2.31.0
scipy>=1.10.0
cartopy>=0.22.0
Pillow>=10.0.0
xarray>=2023.1.0
dask>=2023.1.0
netcdf4>=1.6.0
h5py>=3.8.0
pyproj>=3.5.0
shapely>=2.0.0